lots of things

+++ customer-service/src/main/java/com/customerservice/customerservice/repository/CustomerAddressRepository.java
@@ -3,5 +3,5 @@ package com.customerservice.customerservice.repository;
 import com.customerservice.customerservice.entity.CustomerAddress;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerAddressRepository extends JpaRepository<CustomerAddress, Long> {
+public interface CustomerAddressRepository extends JpaRepository<CustomerAddress, Integer> {
 }
